backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
python-multipart==0.0.20
reportlab==4.4.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
